+++ packages/manager/apps/container/src/helpers/index.ts
@@ -1,3 +1,4 @@
 export { truncate, capitalize, toScreamingSnakeCase } from './stringHelper/stringHelper';
 export { groupBy } from './arrayHelper/arrayHelper';
 export { fromNow } from './dateHelper';
+export { deleteMatchingCookies, deleteCookie } from './cookieHelper';