+++ packages/manager/modules/models/BillingService/BillingService.class.js
@@ -299,6 +299,7 @@ export default class BillingService {
       'SMS',
       'EMAIL_DOMAIN',
       'VEEAM_ENTERPRISE',
+      'HOSTING_WEB_NEW',
     ].includes(this.serviceType);
   }