+++ packages/manager/apps/ips/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.3](https://github.com/ovh/manager/compare/@ovh-ux/manager-ips-app@0.8.2...@ovh-ux/manager-ips-app@0.8.3) (2026-05-20)
+
+
+### Bug Fixes
+
+* **network-ip:** fix order button display and geolocation reset on quantity change ([df81485](https://github.com/ovh/manager/commit/df814857f4e9a26d2a8371d067a2ea1dee0b41a8)), closes [#MANAGER-19813](https://github.com/ovh/manager/issues/MANAGER-19813)
+
+
+
+
+
 ## [0.8.2](https://github.com/ovh/manager/compare/@ovh-ux/manager-ips-app@0.8.1...@ovh-ux/manager-ips-app@0.8.2) (2026-04-22)