+++ packages/manager/apps/pci-savings-plan/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.26.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-savings-plan-app@1.26.0...@ovh-ux/manager-pci-savings-plan-app@1.26.1) (2026-05-19)
+
+
+### Bug Fixes
+
+* **pci-savings-plan:** update format date ([c8712b4](https://github.com/ovh/manager/commit/c8712b40efcc044dc69a58925dde47e30eeacb1b)), closes [#TAPC-6694](https://github.com/ovh/manager/issues/TAPC-6694)
+
+
+
+
+
 # [1.26.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-savings-plan-app@1.25.0...@ovh-ux/manager-pci-savings-plan-app@1.26.0) (2026-04-28)