+++ packages/manager/modules/support/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.46](https://github.com/ovh/manager/compare/@ovh-ux/manager-support@2.1.45...@ovh-ux/manager-support@2.1.46) (2026-06-08)
+
+**Note:** Version bump only for package @ovh-ux/manager-support
+
+
+
+
+
 ## [2.1.45](https://github.com/ovh/manager/compare/@ovh-ux/manager-support@2.1.44...@ovh-ux/manager-support@2.1.45) (2026-02-05)
 
 **Note:** Version bump only for package @ovh-ux/manager-support